+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 11.1.2 - TBD
+# TBD
 ### Changed
 * Upgrade housekeeping to version 1.0.2