+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # TBD
 ### Added
 * Added ability to configure AWS Server Side Encryption for `S3S3Copier` via `copier-options.s3-server-side-encryption` configuration property.
+### Changed
+* Upgrade housekeeping to version 1.0.2
 
 # 11.1.1 - 2018-02-15
 ### Fixed